make service resilient to external and internal IP addresses
in an ideal world this kind of stupid code wouldn't be needed, but my
internet provider modem and router doesn't allow for NAT loopback, so
I can't use the same ip internally and externally. To workaround this issue I
will just hardcode the internal ip for now. Better solutions could be:

1. a config field for the internal IP
2. an external service to get the internal ip (basically use an external proxy that would query a service that I'd expose and would return the internal ip)
3. add an internal DNS that would map the service name to the internal ip
4. change the modem for one that allows for NAT loopback