use JobScheduler to run the notification service in the background
this requires API >= 21 (lollipop). I might add support for lower API
versions in the future.

Note: I didn't use firebase-jobdispatcher-android because it doesn't support
the use of JobScheduler if the system allows for it; it always uses google
play services. It seems to me that the project is vaporware at the moment -
as a data point it only has 69 commits and since 12/2016 there were at most
5 commits