Started using TestScheduler
I also added a test case for a bug when the user is editing an item and a
update comes in. The Observables exercised in this test case touches three
different schedulers: Android.mainThread, Schedulers.io and
Schedulers.computation (implicitly). It is a somewhat complex scenario and
TestScheduler made the bug far easier to reproduce.