implement call to TaskService API when the drag and drop finishes
NOTE: it is almost impossible - at least for me it was impossible - to have
ItemTouchHelper.Callback.clearView method called by Robolectric. The reason
is that this method is only called at the end of an animation after the item
is moved. I tried tons of different things withouth success. In the end I
preferred to override ItemTouchHelper.Callback.onSelectedChange and add the
code to deal with drag and drop end there.